refactor(search-view): extract pure helpers to lib utils
